% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpdist.R
\name{bpdist}
\alias{bpdist}
\title{Distribution of breakpoints}
\usage{
bpdist(topres, ndigits=0)
}
\arguments{
\item{topres}{results from toptrendy() function}

\item{ndigits}{how many digits to be used when rounding. Default is 0 (return integers).}
}
\value{
The function takes significant genes called from 
the toptrendy() function. For any time point, this function calculates how many genes
have breakpoint at this time point. The output is the numbers of genes sorted
by time point
}
\author{
Ning Leng
}

