% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSegRsq.R
\name{fitSegRsq}
\alias{fitSegRsq}
\title{Fit Segmented regression models on a feature/gene}
\usage{
fitSegRsq(Data, T.Vect = NULL, Max.K = 5, Min.Num.In.Seg = 5,
  Pval.Cut = 0.1, Cut.Diff = 0.1, Num.Try = 100, Keep.Fit = FALSE)
}
\arguments{
\item{Data}{matrix of normalized expression measurements. Rows are genes/features 
and columns are samples.}

\item{T.Vect}{a numerical vector indicating the time-points (or the order of samples). 
If it is NULL (default), then the time/order will be assumed to be equaly spaced from 
1:N (N is number of samples).}

\item{Max.K}{maximum number of breakpoints to consider. For each gene, 
trendy will fit Max.K + 1 models containing 0 -> Max.K breakpoints
(1 -> (Max.K+1) segments). The model with the lowest BIC 
value will be selected (unless Force.Rsq == TRUE, see below).}

\item{Min.Num.In.Seg}{minimum number of samples required to be within 
a segment.}

\item{Pval.Cut}{p-value cutoff. If the p-value of a segment is greater than
PvalCut, then the segment will be called as 'no change'.}

\item{Cut.Diff}{Only used if Force.Rsq is set to TRUE. 
 If the difference between the adjusted R^2 from the k + 1 breakpoint
model and the adjusted R^2 from the k breakpoint model is less than Cut.Diff, 
then the optimal number of breakpoints is set as k instead of k + 1.}

\item{Num.Try}{the number of different seeds to try. If all NumTry runs 
fail, then the linear regression (no breakpoints, one segment) model will
be returned.}

\item{Keep.Fit}{whether to report the fitted object (default is FALSE).}
}
\value{
Trend: direction of each sample; -1: down, 0: no change, 1: up
	Slope: fitted slopes, Slope.Trend: sign of fitted slopes, 
	Slope.Pvalue: p value of each segment, Breakpoint: estimated breakpoints, 
	Fitted.Values: fitted values AdjustedR2: adjusted r value of the model
		Fit: fit object
}
\description{
fits segmented regression models
}
\examples{
d1 <- rbind(c(rep(1,50),1:50), c(100:1))
		rownames(d1) <- c("g1","g2")
		fitsegRsq(d1, "g1")
}
\author{
Ning Leng and Rhonda Bacher
}
