% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatResults.R
\name{formatResults}
\alias{formatResults}
\title{Function to format results for saving.}
\usage{
formatResults(Top.Trendy, Feature.Names = NULL)
}
\arguments{
\item{Top.Trendy}{results from topTrendy() function}

\item{Feature.Names}{an optional vector of features 
(if only interested in outputting a subset of features/genes).}
}
\value{
The function will reformat the output from Trendy so
that it can be easily save as a .txt or .csv file.
If Feature.Names is supplied then only the information
for those features/genes is returned.
}
\description{
format data from Trendy which can be saved for later use.
}
\author{
Rhonda Bacher
}
