% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getCounts}
\alias{getCounts}
\title{getCounts}
\usage{
getCounts(DATA)
}
\arguments{
\item{DATA}{An object of class \code{SummarizedExperiment} that contains 
expression data and metadata}
}
\value{
A \code{matrix} which contains the expression data
 where genes/features are in rows and samples are in columns
}
\description{
Convenient helper function to extract the normalized expression
matrix from the SummarizedExperiment
}
\examples{
  m1 <- rbind(c(rep(1,50),1:50), c(100:1))
  ExampleData <- SummarizedExperiment::SummarizedExperiment(assays=list("Counts"=m1))
  myData <- getCounts(ExampleData)
}
