% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_trendy.R
\name{graph_trendy}
\alias{graph_trendy}
\title{Flot Line Charts.}
\usage{
graph_trendy(message, width = NULL, height = NULL)
}
\arguments{
\item{message}{A dataframe.}

\item{width, height}{nada.}
}
\value{
A flot line chart based on Flot.
}
\description{
Create a Flot line chart based on Flot.
}

