% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_shinyData.R
\docType{data}
\name{Orig.Data}
\alias{Orig.Data}
\title{Example dataset for Trendy Shiny App}
\format{list}
\usage{
data(exampleObject_trendyForShiny)
}
\description{
Example of output from running trendy() on 
TrendyExData.RData (see vignette)
}
\examples{
data(exampleObject_trendyForShiny)
}
\keyword{datasets}
