% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeature.R
\name{plotFeature}
\alias{plotFeature}
\title{Plot features of interest}
\usage{
plotFeature(Data, T.Vect = NULL, Feature.Names, Show.Fit = TRUE,
  Trendy.Out = NULL, Cond.Col = NULL, Cond.Col.Sample = NULL,
  Cond.Prefix = "Day", X.Name = "Time", Y.Name = "Normalized Expression")
}
\arguments{
\item{Data}{matrix of normalized expression measurements. Rows are genes/features 
and columns are samples.}

\item{T.Vect}{a numerical vector indicating the time-points (or the order of samples). 
If it is NULL (default), then the time/order will be assumed to be equaly spaced from 
1:N (N is number of samples).}

\item{Feature.Names}{a list of genes or features to plot}

\item{Show.Fit}{whether plot the segmented regression fitting (default is TRUE)}

\item{Trendy.Out}{segmented regression fitting result from running trendy(); i
f Show.Fit is NULL and Trendy.Out = TRUE, then the
segmented regression will be fit for each of the genes and it may take longer to run}

\item{Cond.Col}{color for each condition, names of this vector should be condition names; 
if it is NULL (default), no legend will be generated}

\item{Cond.Col.Sample}{each sample's color. The vector's length should match number of samples. 
if it is NULL, expression will be shown in black (default).}

\item{Cond.Prefix}{name to be used in the condition legend (Default is Day, but could be hour, minute, etc.).}

\item{X.Name}{x-axis name}

\item{Y.Name}{y-axis name}
}
\value{
plot of gene expression and fitted line
}
\description{
plot each feature with (or without) the fitted trend.
}
\examples{
d1 <- rbind(c(rep(1,50),1:50), c(100:1))
	rownames(d1) <- c("g1","g2")
	plotmarker(d1, Feature.Names=c("g1","g2"))
}
\author{
Ning Leng and Rhonda Bacher
}
