% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmarker.R
\name{plotmarker}
\alias{plotmarker}
\title{Plot markers of interest}
\usage{
plotmarker(Data, T.Vect = NULL, File.Name = NULL, Feature.Names, PDF = FALSE,
					Seg.Fit = TRUE, Seg.Data = NULL, Cond.Col = NULL, 
					Cond.Col.Sample = NULL, Cond.Prefix = "Day", 
					Y.Name = "Normalized Expression", Par.Param = c(3,2),
					PDF.Height = 15, PDF.Width = 10)
}
\arguments{
\item{Data}{matrix of normalized expression measurements. Rows are genes 
and columns are samples.}

\item{T.Vect}{a numerical vector indicating the time points. If it is NULL
(default), then the time will be assumed to be equaly spaced from 
1:N (N is number of samples).}

\item{File.Name}{the file name (and file path) to save the Trendy object, 
only used if SaveObject=TRUE.}

\item{Feature.Names}{a list of genes or features to plot}

\item{PDF}{whether to output plots to a pdf file (default is FALSE)}

\item{Seg.Fit}{whether plot the segmented regression fitting (default is TRUE)}

\item{Seg.Data}{segmented regression fitting result from running trendy(); i
f Seg.Data is NULL and Seg.Fit = T, then the
segmented regression will be fit for each of the genes and it may take longer to run}

\item{Cond.Col}{color for each condition, names of this vector should be condition names; 
if it is NULL (default), no legend will be generated}

\item{Cond.Col.Sample}{each sample's color. The vector's length should match number of samples. 
if it is NULL, expression will be shown in black (default).}

\item{Cond.Prefix}{prefix to be appended to condition name (Day is default)}

\item{Y.Name}{y-axis name}

\item{Par.Param}{mfrow for plot layout(Default is c(3,2))}

\item{PDF.Height, PDF.Width}{the height and width for the PDF file}
}
\value{
plot of gene expression and fitted line
}
\description{
plot each feature with (or without) the fitted trend.
}
\examples{
d1 <- rbind(c(rep(1,50),1:50), c(100:1))
	rownames(d1) <- c("g1","g2")
	plotmarker(d1, Feature.Names=c("g1","g2"), PDF = FALSE)
}
\author{
Ning Leng and Rhonda Bacher
}
