% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toptrendy.R
\name{toptrendy}
\alias{toptrendy}
\title{obtain top genes from trendy results}
\usage{
toptrendy(Seg.Data, AdjR.Cut = 0.5)
}
\arguments{
\item{Seg.Data}{output from the trendy function}

\item{AdjR.Cut}{cutoff for the adjusted r^2. Genes whose adjusted
r^2 is greater than adjR.Cut are called as significant.}
}
\value{
only significant genes will be included in the output. 
	The output is reformatted as:
	id.sign direction of each sample; -1: down, 0: no change, 1: up
	slp: fitted slopes, slp.sign: sign of fitted slopes, 
	slp.pval: p value of each segment, 
	bp: estimated breakpoints, fitted: fitted values 
	radj: adjusted r value of the model
	fit: fit object
}
\description{
reformats the list output for genes with a given adjusted R^2 cutoff
}
\examples{
d1 <- rbind(c(rep(1,50),1:50), c(100:1))
     rownames(d1) <- c("g1","g2")
     seg.all <- trendy(d1)
		top.seg <- toptrendy(seg.all)
}
