% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendHeatmap.R
\name{trendHeatmap}
\alias{trendHeatmap}
\title{Draw heatmap of gene expression trends}
\usage{
trendHeatmap(Top.Trendy)
}
\arguments{
\item{Top.Trendy}{results from topTrendy() function}
}
\value{
The function takes significant genes/features called from 
	the topTrendy() function. These genes are further grouped into three
	groups: up, down, or no change in the first segment. Within each group,
	the genes are sorted by their first break point. The heatmap shows expression 
	trends of these three groups of genes. In the heatmap, red/blue/black 
	represents up/down/nochange. A list of genes in the heatmap order is returned.
}
\description{
heatmap of the fitted trends
}
\author{
Ning Leng and Rhonda Bacher
}
