% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendy.R
\name{trendy}
\alias{trendy}
\title{segmented regression on a set of genes}
\usage{
trendy(Data = NULL, T.Vect = NULL, Save.Object = FALSE,
  File.Name = NULL, Mean.Cut = 10, Max.K = 3, Min.Num.In.Seg = 5,
  Pval.Cut = 0.1, Num.Try = 100, Keep.Fit = FALSE, N.Cores = NULL,
  Feature.Names = NULL, Force.Rsq = FALSE, Cut.Diff = 0.1)
}
\arguments{
\item{Data}{matrix of normalized expression measurements. Rows are genes/features 
and columns are samples.}

\item{T.Vect}{a numerical vector indicating the time-points (or the order of samples). 
If it is NULL (default), then the time/order will be assumed to be equaly spaced from 
1:N (N is number of samples).}

\item{Save.Object}{if TRUE then the trendy object produced will be saved 
to use in the shiny app (default is FALSE).}

\item{File.Name}{the file name (and file path) to save the Trendy object, 
only used if SaveObject=TRUE (default name is Data_trendyForShiny.RData).}

\item{Mean.Cut}{genes whose mean is less than MeanCut will not be 
considered, default is 10.}

\item{Max.K}{maximum number of breakpoints to consider. For each gene, 
trendy will fit Max.K + 1 models containing 0 -> Max.K breakpoints
(1 -> (Max.K+1) segments). The model with the lowest BIC 
value will be selected (unless Force.Rsq == TRUE, see below).}

\item{Min.Num.In.Seg}{minimum number of samples required to be within 
a segment.}

\item{Pval.Cut}{p-value cutoff. If the p-value of a segment is greater than
PvalCut, then the segment will be called as 'no change'.}

\item{Num.Try}{the number of different seeds to try. If all NumTry runs 
fail, then the linear regression (no breakpoints, one segment) model will
be returned.}

\item{Keep.Fit}{whether to report the fitted object (default is FALSE).}

\item{N.Cores}{number of cores to use, default is detectCores() - 1.}

\item{Force.Rsq}{whether to use adjusted R^2 in model evaluation 
instead of BIC, corresponds to very early versions of Trendy. Requires that
each succesive increase in adjusted R squared by larger than Cut.Diff.}

\item{Cut.Diff}{Only used if Force.Rsq is set to TRUE. 
 If the difference between the adjusted R^2 from the k + 1 breakpoint
model and the adjusted R^2 from the k breakpoint model is less than Cut.Diff, 
then the optimal number of breakpoints is set as k instead of k + 1.}
}
\value{
Trend: direction of each sample; -1: down, 0: no change, 1: up
	Slope: fitted slopes, Slope.Trend: sign of fitted slopes, 
	Slope.Pvalue: p value of each segment, Breakpoint: estimated breakpoints, 
	Fitted.Values: fitted values AdjustedR2: adjusted r value of the model
		Fit: fit object
}
\description{
Segmented regression models are fit for each gene. 
 The number of model fits is 1 -> Max.K.
}
\examples{
d1 <- rbind(c(rep(1,50),1:50), c(100:1))
     rownames(d1) <- c("g1","g2")
     trendy(d1)
}
\author{
Ning Leng and Rhonda Bacher
}
