% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendyExampleData.R
\docType{data}
\name{TrendyExampleData}
\alias{TrendyExampleData}
\title{Example dataset for Trendy}
\format{data matrix}
\usage{
data(TrendyExampleData)
}
\description{
Example time-course dataset.
}
\examples{
data(TrendyExampleData)
}
\keyword{datasets}
