% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{trendyShiny}
\alias{trendyShiny}
\title{Trendy shiny app to interactively vizualize results after running trendy().}
\value{
Opens a browser window with an interactive \code{shiny} 
 app and visualize all precomputed Trendy fits.
}
\description{
Trendy shiny app to interactively vizualize results after running trendy().
}
