% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointDist.R
\name{breakpointDist}
\alias{breakpointDist}
\title{Distribution of breakpoints}
\usage{
breakpointDist(Top.Trendy, N.Digits = 0)
}
\arguments{
\item{Top.Trendy}{results from topTrendy() function}

\item{N.Digits}{how many digits to be used when rounding 
(default is 0 (return integers))}
}
\value{
The function takes significant genes called from 
the topTrendy() function. For any time point, this function calculates how many genes
have a breakpoint at this time point. The output is the numbers of genes sorted
by time point.
}
\description{
calculates number of breakpoints at each time.
}
\author{
Ning Leng
}
