% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{.breakpointFit}
\alias{.breakpointFit}
\title{break point fits}
\usage{
.breakpointFit(J, tVectIn, lmLinear, numTry)
}
\arguments{
\item{J}{number of breakpoints in the model}
}
\description{
break point fits
}
