% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractpattern.R
\name{extractpattern}
\alias{extractpattern}
\title{Extract pattern from segmented regression}
\usage{
extractpattern(Seg.Data, Pattern = NULL, Radj.Cut = .5, Delay = 0)
}
\arguments{
\item{Seg.Data}{output from trendy() function}

\item{Pattern}{vector containing pattern to search genes/features. If length
is one then it will only consider features with constant pattern 
across the entire time-course.}

\item{Radj.Cut}{only consider features with adjusted R^2 > Radj.Cut. 
Default = .5.}

\item{Delay}{search for pattern starting after certain time-point (e.g. only
genes with a breakpoint > 10).}
}
\value{
genes: names of genes/features containing pattern and the breakpoints
	corresponding to the pattern.
}
\description{
find dynamic genes that follow a given pattern
}
\examples{
myTrends <- trendy(TrendyExData)
extractpattern(myTrends, Pattern = c("up")) #increasing only features
 extractpattern(myTrends, Pattern = c("up", "down")) #features with a peak
 extractpattern(myTrends, Pattern = c("up", "down"), Delay = 20)
}
\author{
Rhonda Bacher
}
