% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSegRsq.R
\name{fitSegRsq}
\alias{fitSegRsq}
\title{Fit Segmented regression models on a feature/gene}
\usage{
fitSegRsq(Data, tVectIn = NULL, maxK = 5, minNumInSeg = 5,
  pvalCut = 0.1, cutDiff = 0.1, numTry = 100, keepFit = FALSE)
}
\arguments{
\item{Data}{a matrix of normalized expression measurements. Rows are 
genes/features and columns are samples.}

\item{tVectIn}{a numerical vector indicating the time-points or the 
order of samples. If it is NULL (default), then the time/order 
will be assumed to be equaly spaced from 1:N (N is number of samples).}

\item{maxK}{maximum number of breakpoints to consider. For each gene,
trendy will fit maxK + 1 models containing 0 -> maxK breakpoints
(1 -> (maxK + 1) segments). The model with the lowest BIC 
value will be selected (unless forceRsq = TRUE, see below).}

\item{minNumInSeg}{minimum number of samples required to be within 
a segment. If a breakpoint model has a segment with fewer than 
minNumInSeg point in any segment, then the model is not 
considered valid.}

\item{pvalCut}{p-value cutoff. If the p-value of a segment is greater than
PvalCut, then the segment will be called as 'no change'.}

\item{cutDiff}{Only used if forceRsq is set to TRUE. 
If the difference between the R^2 from the k + 1 breakpoint
model and the R^2 from the k breakpoint model is less than 
cutDiff, then the optimal number of breakpoints is set as k instead 
of k + 1.}

\item{numTry}{the number of different seeds to try. If all numTry runs 
fail, then the linear regression (no breakpoints, one segment) model 
will be returned.}

\item{keepFit}{whether to report the fitted object (default is FALSE).}
}
\value{
Trend: direction of each sample; -1: down, 0: no change, 1: up
 Slope: fitted slopes, Slope.Trend: sign of fitted slopes, 
 Slope.Pvalue: p value of each segment, Breakpoint: estimated breakpoints,
 Fitted.Values: fitted values AdjustedR2: adjusted r value of the model
 Fit: fit object
}
\description{
fits segmented regression models
}
\author{
Ning Leng and Rhonda Bacher
}
