% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTrendy.R
\name{formatFunc}
\alias{formatFunc}
\title{helper function to format results}
\usage{
formatFunc(IN)
}
\arguments{
\item{IN}{the object to be formatted}
}
\value{
a formated matrix of results
}
\description{
helper function to format result
}
\author{
Rhonda Bacher
}
