% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatresults.R
\name{formatresults}
\alias{formatresults}
\title{Function to format results for saving.}
\usage{
formatresults(Top.Trendy, Feature.Names = NULL)
}
\arguments{
\item{Top.Trendy}{results from toptrendy() function}

\item{Feature.Names}{an optional vector of genes to use.}
}
\value{
The function will reformat the output from Trendy so
that it can be easily save as a .txt or .csv file.
If Feature.Names is supplied then only the information
for those genes is returned.
}
\description{
format data to save.
}
\author{
Rhonda Bacher
}
