% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getCounts}
\alias{getCounts}
\title{getCounts}
\usage{
getCounts(DATA)
}
\arguments{
\item{DATA}{An object of class \code{SummarizedExperiment} that contains 
expression data and metadata}
}
\value{
A \code{matrix} which contains the expression data
 where genes/features are in rows and samples are in columns
}
\description{
Convenient helper function to extract the normalized 
 expression matrix from the SummarizedExperiment
}
\examples{
 m1 <- matrix(c(c(rnorm(50,5,1),sort(rnorm(50, 15, 5))), rnorm(100, 50,10)), 2, 100, TRUE)
 ExampleData <- 
 SummarizedExperiment::SummarizedExperiment(assays=list("Counts"=m1))
 myData <- getCounts(ExampleData)
}
