% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeature.R
\name{plotFeature}
\alias{plotFeature}
\title{Plot features of interest}
\usage{
plotFeature(Data, tVectIn = NULL, featureNames, showFit = TRUE,
  simple = FALSE, showLegend = TRUE, trendyOutData = NULL,
  condCol = NULL, condColSample = NULL, condPrefix = "Day",
  xlab = "Time", ylab = "Normalized Expression")
}
\arguments{
\item{Data}{a matrix of normalized expression measurements. Rows are 
genes/features and columns are samples.}

\item{tVectIn}{a numerical vector indicating the time-points or the 
order of samples. If it is NULL (default), then the time/order 
will be assumed to be equaly spaced from 1:N (N is number of samples).}

\item{featureNames}{a list of genes or features to plot}

\item{showFit}{whether to plot the segmented regression 
fitting (default is TRUE)}

\item{simple}{if TRUE the plot will not highlight the breakpoints and segments 
and will only display a black fitted line. (default is FALSE)}

\item{showLegend}{if TRUE and simple=FALSE then a legend will be output (default = TRUE)}

\item{trendyOutData}{segmented regression fitting result from 
running trendy(); if showFit is TRUE and trendyOutData is NULL, then the
segmented regression will be fit for each of the genes and it may 
take longer to run}

\item{condCol}{color for each condition, names of this vector should be
condition names; if it is NULL (default), no legend will be generated}

\item{condColSample}{each sample's color. The vector's length should match
number of samples. if it is NULL, expression will be shown in 
black (default).}

\item{condPrefix}{name to be used in the condition legend (Default is Day, 
but could be hour, minute, etc.).}

\item{xlab}{x-axis name}

\item{ylab}{y-axis name}
}
\value{
plot of gene expression and fitted line
}
\description{
plot each feature with (or without) the fitted trend.
}
\examples{
d1 <- rbind(c(rep(1,50),seq_len(50)), rev(seq_len(100)))
 rownames(d1) <- c("g1","g2")
 colnames(d1) <- paste0("time", seq_len(100))
 plotFeature(d1, featureNames=c("g1","g2"))
}
\author{
Ning Leng and Rhonda Bacher
}
