% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toptrendy.R
\name{toptrendy}
\alias{toptrendy}
\title{obtain top genes from trendy results}
\usage{
toptrendy{res.trendy, r.cut=.5}
}
\arguments{
\item{res.trendy}{output from the trendy function}

\item{r.cut}{cutoff for the adjusted r square. genes whose adjusted
r square is greater than r.cut are called as significant.}
}
\value{
only significant genes will be included in the output
id.sign direction of each sample; -1: down, 0: no change, 1: up
slp: fitted slopes, slp.sign: sign of fitted slopes, slp.pval: p value of each segment, 
bp: estimated breakpoints, fitted: fitted values radj: adjusted r value of the model
fit: fit object
}

