% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendheatmap.R
\name{trendheatmap}
\alias{trendheatmap}
\title{Draw heatmap of gene expression trends}
\usage{
trendheatmap(topres, showplot)
}
\arguments{
\item{topres}{results from toptrendy() function}

\item{showplot}{whether show the heatmap}
}
\value{
The function takes significant genes called from 
the toptrendy() function. These genes are further grouped into three
groups : go up/down or no change in the first segment. Within each group,
the genes are sorted by their first break point. The heatmap shows expression trends of these 
three groups of genes. In the heatmap, red/blue/black represents up/down/nochange.
}
\author{
Ning Leng
}

