% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendyExampleData.R
\docType{data}
\name{trendyExampleData}
\alias{trendyExampleData}
\title{Example dataset for Trendy}
\format{data matrix}
\usage{
data(trendyExampleData)
}
\description{
Example time-course dataset.
}
\examples{
data(trendyExampleData)
}
\keyword{datasets}
