% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{TrendyExData}
\alias{TrendyExData}
\title{Example dataset for Trendy}
\format{data matrix}
\usage{
data(TrendyExData)
}
\description{
Example time-course dataset.
}
\examples{
data(TrendyExData)
}
\keyword{datasets}
